chore: bump pinned nightly 2026-02-15 -> 2026-07-23
build-std is still required (not stabilizing), so we stay on nightly; this just
refreshes the compiler. Binary 275 KB (static CRT, ~8 KB smaller than the old
nightly), all PE mitigations present, 13 imports (0 network), 74 tests pass.
Applied the fresher nightly's new clippy lints (unnecessary_cast, collapsible_match).
Workflows bumped to match.

+++ app/src/log.rs
@@ -298,7 +298,7 @@ pub fn stack_sample(label: &str) {
     let bin = if kb < 4 {
         0
     } else {
-        let log2 = usize::BITS - 1 - kb.leading_zeros() as u32;
+        let log2 = usize::BITS - 1 - kb.leading_zeros();
         (log2 as usize).saturating_sub(1).min(STACK_NUM_BINS - 1)
     };
     STACK_BINS[bin].fetch_add(1, Ordering::Relaxed);
